refactor(react): extract secret-backed header auth UI
